Fix EventStore circular dep
